Update to dbinit.sql
According to pgAdmin III, the client I'm using, this is accurate. The code executes--I had to learn to put the tables in the right order based on the dependencies. I also checked the execution in order to make sure the primary keys and foreign keys were properly assigned. According to pgAdmin III, they are. Let me know if any of this turns out to be incorrect, after all.